+++ prisma/migrations/20260424145859_add_allow_reschedule/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "public"."ReservationSettings" ADD COLUMN     "allowCustomerReschedule" BOOLEAN NOT NULL DEFAULT true;